Initial preparation for Virtual Universe 1.0.2 RC2
** NOTE ** Grid users should run WhiteCore.Server initially to complete this migration before restarting normally ***
NOTE: Version 1.0.2 RC2 will officially be out on April 30, 2016

- Merge in updates from our friends at WhiteCore (many thanks you rock!)
- Code refactoring
- Corrections to incorrect code as needed
- Implements GroupCurrency and updates necessary connectors and database tables
- Updates for CAPS
- Increases the timeout in WebUtils from 5000ms to 10000ms to allow VIrtual Universe to play nice with older hardware (should solve the constant slow messaging posts in consoles)
- Default to using 'FileBased' assets instead of SQL database (grid/standalone)
- Enable 'FileBasedAssets' for Standalone instances
- Head rotated log with current uptime details
- Adopt Greythane's proposed changes to logging
  -- Specify log file name rather than trying to get from OS which only successfully works on Windows
  -- Grid server  - WhiteCore_Grid_yyyymmdd.log
  -- Sim/Region - WhiteCore_Sim_yyyymmdd.log
- General cleanup
- Licensing fixes

+++ Universe/Services/DataService/Connectors/Local/LocalUserInfoConnector.cs
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) Contributors, http://virtual-planets.org/,  http://whitecore-sim.org/, http://aurora-sim.org
+ * Copyright (c) Contributors, http://virtual-planets.org/, http://whitecore-sim.org/, http://aurora-sim.org
  * See CONTRIBUTORS.TXT for a full list of copyright holders.
  *
  * Redistribution and use in source and binary forms, with or without
@@ -224,7 +224,7 @@ namespace Universe.Services.DataService
         public uint RecentlyOnline(uint secondsAgo, bool stillOnline)
         {
              //Beware!! login times are UTC!
-            int now = Util.ToUnixTime(DateTime.Now.ToUniversalTime()) - (int) secondsAgo;
+            int now = Util.ToUnixTime (DateTime.Now.ToUniversalTime ()) - (int)secondsAgo;
 
             QueryFilter filter = new QueryFilter();
             filter.orGreaterThanEqFilters["LastLogin"] = now;
@@ -245,7 +245,7 @@ namespace Universe.Services.DataService
             if (secondsAgo > 0)
             {
                 //Beware!! login times are UTC!
-                int now = Util.ToUnixTime(DateTime.Now.ToUniversalTime()) - (int) secondsAgo;
+                int now = Util.ToUnixTime (DateTime.Now.ToUniversalTime ()) - (int)secondsAgo;
 
                 filter.orGreaterThanEqFilters ["LastLogin"] = now;
                 filter.orGreaterThanEqFilters ["LastSeen"] = now;
@@ -263,14 +263,14 @@ namespace Universe.Services.DataService
                                              uint start, uint count)
         {
             //Beware!! login times are UTC!
-            int now = Util.ToUnixTime(DateTime.Now.ToUniversalTime()) - (int) secondsAgo;
+            int now = Util.ToUnixTime (DateTime.Now.ToUniversalTime ()) - (int)secondsAgo;
 
             QueryFilter filter = new QueryFilter();
             filter.orGreaterThanEqFilters["LastLogin"] = now;
             filter.orGreaterThanEqFilters["LastSeen"] = now;
             if (stillOnline)
             {
-                //filter.andGreaterThanFilters["LastLogout"] = now;
+//                filter.andGreaterThanFilters["LastLogout"] = now;
                 filter.andFilters["IsOnline"] = "1";
             }
 
@@ -287,7 +287,7 @@ namespace Universe.Services.DataService
             if (secondsAgo > 0)
             {
                 //Beware!! login times are UTC!
-                int now = (int)Util.ToUnixTime (DateTime.Now.ToUniversalTime ());
+                int now = Util.ToUnixTime (DateTime.Now.ToUniversalTime ());
                 now -= (int) secondsAgo;
 
                 filter.orGreaterThanEqFilters["LastLogin"] = now;
@@ -296,6 +296,7 @@ namespace Universe.Services.DataService
 
             // online only please...
             filter.andFilters["IsOnline"] = "1";
+
 
             List<string> query = GD.Query(new string[] { "*" }, m_userInfoTable, filter, sort, start, count);
